requirements.txt
# Core Flask and Web Framework
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.5
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Cors==4.0.0
Flask-Security-Too==5.3.0
Flask-JWT-Extended==4.5.3

# Database and ORM
psycopg2-binary==2.9.9
SQLAlchemy==2.0.21

# Task Queue and Caching
redis==5.0.1
celery==5.3.4
celery-redbeat==2.1.1

# Security and Authentication
PyJWT==2.8.0
bleach==6.1.0
html5lib==1.1
flask-talisman==1.1.0
cryptography==41.0.7
bcrypt==4.1.2

# Azure SDK
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-resource==23.0.1
azure-monitor-opentelemetry==1.0.0
azure-mgmt-chaos==1.0.0b2
msal==1.26.0

# Data Processing and ML
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
joblib==1.3.2
scipy==1.11.4

# OpenTelemetry and Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-exporter-otlp==1.21.0
opentelemetry-propagator-b3==1.21.0
structlog==23.2.0
prometheus-client==0.19.0

# Threat Intelligence and Security Tools
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-stix2==3.0.1
cybox==2.1.0.21
mixbox==1.0.5

# API and Validation
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0
apispec==6.3.0
webargs==8.3.0

# Monitoring and Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1
prometheus-client==0.19.0
sentry-sdk==1.32.0

# Rate Limiting and Performance
Flask-Limiter==3.5.0
gunicorn==21.2.0
psutil==5.9.5

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
python-dateutil==2.8.2
pytz==2023.3
urllib3==2.1.0

# SOAR Integration
pycortex==2.0.0
pymisp==2.4.176

# Chaos Engineering
chaostoolkit==1.17.0
chaostoolkit-azure==0.12.0

# Compliance and Governance
policy-sentry==0.12.10